requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
peewee==3.17.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
groq==0.4.1
reportlab==4.0.7
qrcode[pil]==7.4.2
aiofiles==23.2.1
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-decouple==3.8
pdf2image==1.16.3
Pillow==10.1.0